pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dag-factory"
dynamic = ["version"]
description = "Dynamically build Apache Airflow DAGs from YAML files"
authors = [{ name = "Astronomer", email = "humans@astronomer.io" }]
readme = "README.md"
license = "Apache-2.0"
license-files = { paths = ["LICENSE"] }
requires-python = ">=3.8"
keywords = ["airflow", "apache-airflow", "provider", "astronomer", "dag"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Database",
    "Framework :: Apache Airflow",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "apache-airflow>=2.3",
    "apache-airflow-providers-http>=2.0.0",
    "apache-airflow-providers-cncf-kubernetes",
    "pyyaml",
    "packaging",
    "astor>=0.8",
]


[project.optional-dependencies]
tests = [
    "apache-airflow-providers-slack",
    "pytest>=6.0",
    "pytest-cov",
    "pre-commit",
]

######################################
# TESTING
######################################

[tool.hatch.envs.tests]
dependencies = [
    "dag-factory[tests]",
    "apache-airflow~={matrix:airflow}.0,!=2.9.0,!=2.9.1",  # https://github.com/apache/airflow/pull/39670
    "httpx>=0.25.0",
    "pandas",
]
pre-install-commands = ["sh scripts/test/pre-install-airflow.sh {matrix:airflow} {matrix:python}"]

[[tool.hatch.envs.tests.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]
airflow = ["2.3", "2.4", "2.5", "2.6", "2.7", "2.8", "2.9", "2.10"]


[tool.hatch.envs.tests.scripts]
freeze = "pip freeze"
static-check = " pre-commit run --files dagfactory/*"
test = 'sh scripts/test/unit.sh'
test-cov = 'sh scripts/test/unit-cov.sh'
test-integration = 'sh scripts/test/integration.sh'
test-integration-setup = 'sh scripts/test/integration-setup.sh'

[project.urls]
Source = "https://github.com/astronomer/dag-factory"

[tool.hatch.version]
path = "dagfactory/__init__.py"

[project.entry-points."airflow.plugins"]
dagfactory = "dagfactory.plugin:DagFactoryPlugin"

[tool.hatch.build]
sources = ["."]

[tool.hatch.build.targets.sdist]
include = ["dagfactory"]

[tool.hatch.build.targets.wheel]
packages = ["dagfactory"]

[tool.distutils.bdist_wheel]
universal = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
minversion = "6.0"
markers = ["integration", "callbacks"]

######################################
# DOCS
######################################

[tool.hatch.envs.docs]
dependencies = [
    "mkdocs",
    "mike",
    "pymdown-extensions",
    "mkdocs-material",
]

[tool.hatch.envs.docs.scripts]
dev = "mkdocs build && mkdocs serve" # For local development and preventing publishing
gh-deploy = "python scripts/docs_deploy.py dev"
gh-release = "python scripts/docs_deploy.py release"

######################################
# THIRD PARTY TOOLS
######################################

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["C901", "D300", "I", "F"]
ignore = ["F541", "C901"]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["dagfactory", "tests"]

[tool.ruff.lint.mccabe]
max-complexity = 10
